# swagger.yaml
openapi: 3.0.0
info:
  title: 4TP API
  description: 4TP API
  version: 0.0.1
servers:
  - url: "https://for-the-players.herokuapp.com/"
...
paths:
...
  /platforms:
    post:
      summary: Create a new platform
      security:
        - BearerAuth: [ ]
      tags:
        - platform
      operationId: createPlatform
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  description: The new platform's name
                  type: string
                show:
                  description: If the platform should be showed to the end user
                  type: boolean
                imgur:
                  description: >
                    The imgur link of the new platform's image.
                    If this is present `imgBase64` will be ignored
                  type: string
                imgBase64:
                  description: >
                    The base64 of the new platform's image.
                    If `imgur` is present this will be ignored
                  type: string
      responses:
        "400":
          description: "The input data was wrongly formatted"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemaValidationError"
        "401":
          description: "Missing bearer token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "201":
          description: "Platform successfully added"
    get:
      summary: Returns the list of all platforms that can be seen by the end user (The one that have visible=true)
      security:
        - BearerAuth: [ ]
      tags:
        - platform
      operationId: listAllVisiblePlatforms
      responses:
        "401":
          description: "Missing bearer token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "200":
          description: "List of platforms"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    _id:
                      type: string
                      description: The platform's id
                    name:
                      type: string
                      description: The platform's name
                    imageUrl:
                      type: string
                      description: The URL of the image that represents this platform
                    games:
                      type: array
                      items:
                        type: object
                        properties:
                          _id:
                            type: string
                            description: The game's ID
                          name:
                            type: string
                            description: The game's name
  /platforms/{platformId}:
    parameters:
      - name: platformId
        in: path
        required: true
        description: The MongoDB ID of the platform
        schema:
          $ref: "#/components/schemas/MongoID"
    get:
      tags:
        - platform
      security:
        - BearerAuth: [ ]
      summary: Returns the platforms that corresponds to the platform ID
      responses:
        "401":
          description: "Missing bearer token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "404":
          description: "Platform not found"
        "200":
          description: "The platform"
          content:
            application/json:
              schema:
                type: object
                properties:
                  name:
                    description: The platform's name
                    type: string
                  show:
                    description: If the platform should be showed to the end user
                    type: boolean
                  imageUrl:
                    type: string
                    description: The URL of the image that represents this platform
                  games:
                    type: array
                    items:
                      type: object
                      properties:
                        _id:
                          type: string
                          maxLength: 24
                          minLength: 24
                          description: The mongo ID of the game
                        name:
                          type: string
                          description: The name's game
                        show:
                          type: boolean
                          description: If the user should see this game for the platform
...
  /rulesets:
    get:
      summary: Returns all rulesets
      security:
        - BearerAuth: [ ]
      tags:
        - rulesets
      operationId: listRulesets
      responses:
        "400":
          description: "The input data was wrongly formatted"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemaValidationError"
        "401":
          description: "Missing bearer token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "200":
          description: "The list of all rulesets"
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    name:
                      type: string
                      description: "The name of the ruleset"
                    _id:
                      type: string
                      description: The Mongo ID of the ruleset
                    game:
                      type: object
                      properties:
                        _id:
                          type: string
                          description: The Mongo ID of the game the ruleset is for
                        name:
                          type: string
                          description: The name of the game the ruleset is for
                    description:
                      type: string
                      description: The description of the ruleset
                    maps:
                      type: array
                      description: The maps that can be played in this tournament
                      items:
                        type: string
                    bestOf:
                      description: If the value is "3" indicates that when this ruleset is used matches are played in a best of 3
                      type: integer
                    maxNumberOfPlayersPerTeam:
                      type: integer
                      description: The maximum number of players that can be in a team when using this ruleset
                    minNumberOfPlayersPerTeam:
                      type: integer
                      description: The minimum number of players that can be in a team when using this ruleset
    post:
      summary: Create a new ruleset for a game
      security:
        - BearerAuth: [ ]
      tags:
        - rulesets
      operationId: createRuleset
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required:
                - game
                - maxNumberOfPlayersPerTeam
                - minNumberOfPlayersPerTeam
                - description
                - name
                - maps
              properties:
                name:
                  description: The ruleset's name, must be unique
                  type: string
                maxNumberOfPlayersPerTeam:
                  description: The maximum number of players a team can have
                  type: integer
                minNumberOfPlayersPerTeam:
                  description: The minimum number of players a team can have
                  type: integer
                description:
                  description: The ruleset's description
                  type: string
                game:
                  description: Mongo ID of the game
                  type: string
                  minLength: 24
                  maxLength: 24
                maps:
                  type: array
                  description: The maps that can be played in tournaments using this ruleset
                  items:
                    type: string
                bestOf:
                  description: If the value is "3" indicates that when this ruleset is used matches are played in a best of 3
                  type: integer
      responses:
        "400":
          description: "The input data was wrongly formatted"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SchemaValidationError"
        "401":
          description: "Missing bearer token"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "403":
          description: "Forbidden"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorMessageResponse"
        "201":
          description: "Ruleset added"
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
  schemas:
    SchemaValidationError:
      type: array
      items:
        type: object
        properties:
          value:
            type: string
            description: "The value that was received by the server"
          msg:
            type: string
            description: "Error message"
          param:
            type: string
            description: "Name of the field that caused the error"
          location:
            type: string
            description: "Where the field was that was received by the server"
    ErrorMessageResponse:
      type: object
      properties:
        errorMessage:
          type: string
...
    MongoID:
      type: string
      maxLength: 24
      minLength: 24